Add Destination
Write a unit test for the Test_AddDestination() method.
To create a new Destination, you need to create a valid object with name, location, description, bestTimeToVisit, attractions and category. The category property must contain a valid id of one of the existing categories. To do that you need to retrieve the list of categories and set the category property to have the value of the id of one of the categories returned.
Conditions to Test:
•	Get all categories

•	Create a new Destination

•	Response Assertions:
o	The HTTP response status code should be 200 OK.
o	The response content should not be empty.
•	Retrieve the id of the created Destination from the response

•	Get the details of the Destination

•	Response Assertions:
o	The HTTP response status code should be 200 OK.
o	The response content should not be empty.

•	Destination Fields Assertions:
o	The name in the response should match the input value.
o	The location in the response should match the input value.
o	The description in the response should match the input value.
o	The bestTimeToVisit in the response should match the input value.
o	The category should not be empty.
*	The id of the category should match the input value (keep in mind that here the category is returned as an object, which has the _id property you need).
o	The attractions should be a JSON array.
*	The array should have the same number of elements as the input value for attractions.
*	The values of the elements should be the same as the input values for attractions.